refactor: split compatibility transports by ownership
